Allow restoring subset of local variables.
Because the set of local variables is only known
after compiling a function body, and that the
function body can contain calls to rest_loc_var,
the rest_loc_var calls only contained the
synthetic local variables used so far in the
function body. This was a problem when returning
early from a function, as rest_loc_var would leave
part of the stack frame.
To solve this issue, we would either have had to
fixup the calls to rest_loc_var after the function
body was compiled, or adapt rest_loc_var to work
when not all local variables are being restored.
The latter was chosen, as it is more simple and
faster when not all local variables have been
written to.
The only downside is that stack frames now has an
extra entry to store the number of local
variables.