Add HANDLE_SIMPLE_PRINTF optimization
When activated, it detects printf("...") calls and
compiles them to shell's printf calls. This saves
many defstr calls, which in turn reduce memory
usage and packing/unpacking of strings. Getting
rid of defstr calls makes functions printing
many hardcoded strings, such as prologue, shorter
and easier to read.

However, the difference in bootstrapping time is not
measurable and so it is left disabled.