Partial support for backslash-newlines
TCC only uses backslash newlines in macro
definitions, so the tokenizer now recognizes this
pattern and ignores it.

This is a partial implementation, as it does not
support backslash-newlines in other contexts, such
as in the middle of identifiers.

For example, the following declaration is valid C90
code but is not supported by pnut.

```
defi\
ne FO\
O 10\
20
```

See https://gcc.gnu.org/onlinedocs/gcc-2.95.3/cpp_1.html#SEC2