Generalize labels to support gotos
Like the labels used for control flow in the exe
backend, gotos can either jump forward or
backward. When jumping forward, we don't know the
destination yet and so we need a way to patch-up
the code when the destination is set. This commit
generalizes the label mechanism so it can be used
for gotos.