Fix switch statements blocks with only 1 statement
Because case and default are like AST label nodes,
where there is a single statement trailing them,
that statement was processed differently from the
other statements of the conditional block. Because
we need to check for break/return statements to
determine the end of the block and that check
wasn't applied to that single statement, the
code was not correctly handling the case where
there was only one statement in the block.

Example of bad code:
```c
  int i = 0;
  switch (i) {
    case 1:
      return 0;
    default:
      return i;
  }
```

The code was refactored to use a do-while loop
instead of a while loop, and the first iteration
is done on that trailing statement.