Include C code along with generated Shell code
This feature is behind the INCLUDE_C_CODE flag since we only want to use
it with the Shell backend and for certain artifacts.

The C code has the following transformations applied to it:

1. Preprocessor directives are removed (but uses of macros are left untouched)
2. Extra newlines are removed to make code more compact
3. A header is added to function and enums declarations

With INCLUDE_C_CODE, pnut.sh has 8880 lines, and 5002 lines without.

For ksh, the Shell bootstrap takes a few seconds more:
- Without INCLUDE_C_CODE: 20.8s
- With INCLUDE_C_CODE: 23.1s