Simplify memory allocation of string unpacking
Instead of growing the object each time a character is unpacked, an
object is preallocated based on the length of the shell string. For
escaped strings, this can result in some waste of the address space
since the allocated object may be too large by a factor of 2, but the
extra memory cells aren't used so it doesn't slow down the env.

It simplifies the runtime library, and should slightly speed up string
unpacking.