Replace __INIT_GLOBALS with compile time option
Use RT_NO_INIT_GLOBALS to disable the initialization of array global
variables at runtime. This significantly speeds up pnut-sh, but can
break certain C programs that read memory before it is initialized
because of the `-u` shell option.

The original C semantic could be restored by removing the `-u` option,
but this option is very useful for debugging so it is better to keep it.