Inline putchar and exit primitives
Putchar can be inlined either in the runtime library or in the compiled
code. Inlining putchar hurts readability since it's compiled to a printf
that can be hard to read, but gives better performance and shorter code.

Inlining putchar in the runtime library also exhibit the same compromise
but may be more acceptable since it's only done once in put_pstr.