Read #directives until '\n' token
Because the preprocessor treats whitespace as important, it used to
parse macros by looking for certain characters. That was a problem
when the preprocessor encountered whitespace and comments, and prevented
it from recognizing the end of a macro.

This commit fixes that issue by adding a flag (skip_newlines) indicating
to the tokenizer if it should skip '\n' or not. When treating
preoprocessor directives, this flag is set to false so that the macro
parser can stop at the end of the line.