Make free a no-op for bash-2.05a compatibility
The earliest bash version with which Pnut can bootstrap itself is 2.05b.
Version 2.05a fails when pnut gets to the end of an #included file
because free is called and the `unset` builtin fails (and then pnut
because of set -e). We already have an option that toggle free between a
no-op and using unset, so let's disable it since it doesn't make much of
a difference for pnut.

From bash's `CHANGES` file:

  h.  The `unset' builtin no longer returns a failure status when asked to
      unset a previously-unset variable or function.