Remove int_to_char lookup table
int->char conversion are not very common and can often be avoided. In
the few cases where they are not (when opening a file and when writing
bindings for utilities such as in posix-utils.sh), it is because the C
string needs to be passed to a shell function or utility, and the strings
in those cases are usually small. It may also be faster to use
$(put_pstr $str_ptr) to convert the string since that skips the lookup
process all together (by using the `printf \\000` trick) and only uses 1
subshell for the whole string.

For the bootstrap process, int->char conversions are not needed at all
(except when opening a file), and so removing the lookup table pays for
itself in terms of time saved by not having to compile it.

On bootstrap-pnut.sh, it saves 0.3s for ksh, 1.6s for dash and 1.5s for
bash.

In terms of number of lines, this commit removes 106 lines (1046 => 940)
from sh-runtime.c and 202 lines from pnut.sh (7265 => 7063).