Remove pack_string function
Unlike the string unpacking operation, packing is only required when
a C string must be passed a shell command. While bootstrapping pnut,
this only happens when opening a file, which happens only a few times
and is not performance critical.

It is much simpler to reuse the `put_pstr` function, which is already
used to output strings, and capture its output. This may be faster on
certain shells as it only creates a single subshell for the whole string
instead of 1 per character.

One downside is that `pack_string` was used in `posix-utils.sh` which
may suffer from a slight performance hit. However, a faster `pack_string`
could be implemented directly in `posix-utils.sh` if needed. For now,
`posix-utils.sh` includes a `pack_string` function that uses `put_pstr`.

In terms of number of lines, this commit removes 20 lines (918 => 898)
from sh-runtime.c and 52 lines from pnut.sh (7034 => 6982).