Reuse C parser to parse #if expressions
Now that the tokenizer can produce NEWLINE tokens, we can reuse the C
parser to parse #if expressions. Without newlines, the C parser would
keep reading until the end of the expression, skipping over the newlines.
Now, if it encounters a newline, a newline token is produced and the
C parser fails to parse the expression.

This replaces the code that implemented the shunting yard algorithm
with a function that can evaluate constant expressions. This function
evaluates AST nodes that represent constant expressions, and will be
used to support non-integer literal expressions for array lengths.