Replace almost all wrap_str with wrap_str_const
Almost all calls to wrap_str were with string literals. wrap_str would
allocate a new text object and copy the content of the string literal
into it. That wasted time and memory, which can now be avoided by using
wrap_str_const instead.

On ksh, this save almost 1.5s on `./bootstrap-pnut.sh --fast`.

Looking at the allocation numbers, we can see that max_text_alloc is
about 25% lower and the cumulative text allocation is about 30% lower.
This is a very good improvement in memory usage since the text objects
are the most memory-consuming objects in pnut.

This commit:
  string_pool_alloc=44171 heap_alloc=106896 max_text_alloc=14240 cumul_text_alloc=199883

Main:
  string_pool_alloc=44159 heap_alloc=106847 max_text_alloc=18263 cumul_text_alloc=280550