Add pnut mode to do only expansion of #include
Using the `DEBUG_EXPAND_INCLUDES` option makes Pnut read the input file,
tokenize it (reading files included in the process), and then output
the characters as they are read.

This gives a very accurate view of the total lines that are read by Pnut
and allows us to compute the size of pnut.c for different backends.