More correct empty-block detection logic
The old logic was very ad-hoc and was located with break/continue/return
statements since those are the statements that are typically omitted.
When writing it, I didn't take into account is that compound statements
can be empty, in which case a ':' also needs to be emitted.

The new logic moves the responsability of emitting the ':' to the code
creating the if/loop/function which simplifies the code and fixes the
issue with compound statements.