Track loops and switch in local scope
This removes the need to save and restore start/end loop actions which
were easy to forget. This commit also adds a new case to the
optimization of return statements and returning from a nested function
in tail position now omits the full return protocol and use `break N`
instead.