build:  default to 'sh' rather than login shell
The login shell can be anything.  It may not be a POSIX shell (e.g., it
could be fish) in which case the build would break.  Or it could be zsh,
which would make the build unnecessarily slow.

Default to `sh` instead which is a posix shell.  Add Makefile variable
BUILD_SHELL to make this configurable.