Fix ./run-pnut-variant.sh --tokenizer
In https://github.com/udem-dlteam/pnut/pull/106, the `val` associated
with a `STRING` token was changed from the index in the string pool to a
reference to the hash table probe. I forgot to change the code in
debug.c which expected the old behavior.

This commit fixes the issue by updating the code in debug.c, supporting
at the same time STRING tokens containing NUL characters.