Install time utility in CI jobs setup
On GitHub, the ubuntu-latest image used for running the CI jobs includes
the `time` utility. However, when running the CI jobs locally using act,
the ubuntu-latest image does not include the `time` utility and so it
needs to be installed.

This commit also changes bootstrap-pnut-sh to use /usr/bin/time for
consistency with bootstrap-pnut-exe.