Extend declaration parser to support most cases
The previous parser only handled relatively basic
types (int, pointer to int, arrays of ints) as
well as function declarations.

This new parser handles all type, storage-class
specifiers and type qualifiers, and is recursive,
meaning it is now possible to have declarations
such as `void test(int (*)(int, int))`.

This will allow us to add support for unsigned and
floating pointer types, as well as function
pointers.

The only part of declarations that are left
untouched by this commit is the initializer
parser. This will be done in a follow-up PR.