Parse `sizeof (expr)` correctly
When looking at the C grammar, it looks like sizeof has 2 production
rules (`sizeof (type)` and `sizeof expr`) that can be distinguished by
the presence of parenthesis. When a parenthesis is present, it must be
followed by a type, and no parenthesis means it must be followed by an
expression.

However, this is not true since expressions can be parenthesised meaning
that the 2 rules can be differentiated by considering if the token
that follows the parenthesis is part of a type.