Make storage class specifier parsing more robust
The previous implementation permitted multiple
storage class specifiers at the same time,
function declarations with a leading typedef (that
was ignored), and associated the storage class
specifier with the type specifier instead of the
declarator, which caused issue because the typedef
specifier was saved in the types table.

This new implementation is stricter, and allows
the use of a more compact representation for type
qualifiers (const and volatile).