Convert oct/hex literals to decimal outside $(( ))
In the generated shell code we often assume that
numbers have a canonical representation, for
example when comparing numbers or when matching on
a case constant in a switch statement. By wrapping
the octal and hexadecimal literals in $(( )), we
prevent them from "escaping" into variables and in
contexts where decimal numbers are expected.