Remove --fast opt in bstrp-pnut-exe ksh CI check
When the --fast option is used, it generates a file that sometimes fails
to parse on certain versions of ksh. On the macOS version of ksh that
comes from brew, the following error is generated:

> ksh --version
  version         sh (AT&T Research) 93u+m/1.0.10 2024-08-01
> ksh build/pnut-sh.sh
build/pnut-sh.sh[7759]: readonly: : invalid variable name

From playing with the failing script, I found that seemingly unrelated
changes (changing the order of the readonly variables for example) fixed
the error so the error isn't caused by the --fast option, but removing
--fast gets rid of the error.

Fortunately, ksh is fast enough that the --fast option is not necessary
to make CI complete in a reasonable amount of time.