Automatically stub undefined func with rt error
When UNDEFINED_LABELS_ARE_RUNTIME_ERRORS is
defined, pnut will generate a stub for missing
functions that prints the name of the function and
then return. The plan is to crash the program
after printing the message so we have a good idea
of the missing functions that are actually called.