Use DX instead of BP for x86 temporary register
BP cannot be used as the source register of
mov_mem8_reg as the lbp (low byte of BP) cannot be
encoded in the ModR/M byte. In fact, in 32-bit
mode, only the 2 low byte (0-7 and 8-15 bits)
registers of AX, CX, DX, and BX can be used.

This commit fixes the issue by using DX instead of
BP as the temporary register as it can be used as
the source register of mov_mem8_reg and for other
instructions with restrictions on the operands.