Simplify main function epilogue
Before this commit, there existed multiple
versions of the main function epilogue, depending
on whether it returned an exit code and if it used
the argc/argv arguments. With this commit, the
main function epilogue now has 2 versions:
- one that uses the argc/argv arguments
- one that does not use the argc/argv arguments

The argc/argv arguments part makes sense as
converting from the shell argv to the C argv pulls
a lot of code from the runtime, but the exit code
part doesn't save much code and makes the sh
backend slightly more complex.