Remove IDENTIFIER_EMPTY internal node
IDENTIFIER_EMPTY nodes are only used to signal
that the result of a function is not used. Since
this is only relevant in comp_fun_call_code for
assign_to, we can instead use 0 instead. This
avoids the allocation of 1464 IDENTIFIER_EMPTY
nodes when bootstrapping pnut-sh, in addition to
removing redundant matches on IDENTIFIER_EMPTY
nodes.