Place built-in functions at the start of the code
By declaring and defining the built-in functions
at the start of the code, compiling calls to these
functions will be easier for the compiler since
the addresses of the functions will be known
when the calls are compiled.

The motivation behind this change is to make
pnut-exe dump the machine code it generates
as soon as possible, and calls to functions
that are not yet defined will go through a forward
jump table, which is slower and takes more space.