Fix file/VM size in mach-o header
The file/VM size didn't take into account the size
of the mach-o header, and was off by 360 bytes.
This meant that the last 360 bytes of the program
were not loaded in memory, and were instead
replaced with 0x00 (addb %al, (%rax)).

Because the builtin libc used to be at the end of
the program, as long as the built-in libc
functions were not used, the invalid code would
not execute and the program would work. Now that
the built-in libc functions are at the beginning
of the program, the last instructions are the
globals initialization, which result in a crash.

This bug was not caught before because the last
built-in function was printf, which was stubbed to
print an error message and crash, and so it was
never called. It occupied around 1700 bytes,
which meant all the other built-in functions were
valid.