Shorten code generated by rt_debug
Before, it unrolled the string and for each
character it generated code that did the write
syscall. We can instead call the write syscall
with the whole string at once.

This saves around 2000 bytes of code only for the
rt_crash("printf is not supported yet.") in the
printf stub.