Set cgc_fs = 1 before compiling first function
cgc_fs is a global variable counting the number of
bindings in the local environment. Normally, it is
initialized to 1 since the return variable is
always added implicitely.

Because cgc_fs was set to 1 after compiling a
top-level declaration to reset the counter for the
next declaration, the first function of a program
would be compiled with cgc_fs set to 0, offsetting
the rest of the parameters by one.

This bug only revealed itself when using the
constant parameter optimization, as the code that
initialized variables with positional parameters
used a different counter.