Add empty getcwd implementation
Unfortunately, UNDEFINED_LABELS_ARE_RUNTIME_ERRORS
is not compatible with the one-pass generator
since undefined functions are always called
indirectly and their jump table entry is not
initialized. The solution should be to have
assert_all_labels_defined initialize the jump
table entry for labels that are function entry
points, but we don't have that information at the
moment.

Fortunately, getcwd is the only function that was
missing to make the one-pass generator work with
TCC.