Workaround: Change c1 and c2 types to char
M2-Planet generates code that results in the upper
bits of the c1 and c2 variables not being zero
for `c1 = char_array[index]`.
This results in interned tokens not being found,
breaking the parser. This bug can be worked around
by changing the type of c1 and c2 to char.