Centralize error message formatting
Because pnut-exe does not support printf (runtime
error), the codebase was littered with unrolled
printf calls. Because pnut-sh places each function
call on a separate line, this resulted in noisy
error handling code in pnut-sh.sh and pnut-exe.sh.

The solution is to replace the unrolled printf
calls with a few functions that format error
messages. By doing so, these functions can be
implemented using printf when compiling with
a compiler supporting it, and unrolled when
compiling with pnut-exe.

Because detailled error messages are not very
useful when bootstrapping a known compiler, the
error message formatting functions can also be
macroed out to produce minimal error messages
when bootstrapping.

Line reduction (before):
  pnut-sh.sh: 7755
  pnut-exe.sh: 7294

Line reduction (with detailled error messages):
  pnut-sh.sh: 7744
  pnut-exe.sh: 7272

Line reduction (without detailled error messages):
  pnut-sh.sh: 7625
  pnut-exe.sh: 7150