Clarify error message when taking address of var
The address of operator (&) on variables (local or global) is not
supported in pnut-sh. The error message was previously mentioning "local
variable" even when the variable was global, which was confusing.

To add support for `&variable` would require the following changes:
- A more complete environment tracking, such that global variables can
  be distinguished from other global-level values such as enum cases.
- "Boxing" all variables, so that they have an addressable location.
  This can be done relatively easily for global variables, but local
  variables don't live on a stack which makes it more complex.

Not supporting `&variable` may be revisited in the future, but for now
this change simply clarifies the error message to indicate that taking
the address of any variable is not supported.