Skip empty statements in compound statement parser
pnut-sh optimizes return statements that appear in
tail position, with tail position being defined as
the last statement in a compound statement. But,
if the last statement is an empty statement, i.e.,
a semicolon, then the optimization is not applied.

This commit changes the compound statement parser
to skip empty statements, so that a return
statement followed by a semicolon is still
considered to be in tail position.

This commit reduces the number of lines in
pnut-sh.sh by 5 lines, 4 of which were redundant
endlet statements.