Handle strings with \0 in symbol table
| Version              | now      |
| -------------------- | -------- |
| pnut-sh.sh           | 6221 LOC |
| pnut-exe.sh          | 5618 LOC |
| pnut-exe-one-pass.sh | 5664 LOC |

Before, the symbol table compared strings until a
null character, which made it possible for 2
different strings to be considered identical if
their difference was only after a null character.
This bug didn't happen in practice because such
strings are not used in pnut, and because it would
only happen on strings falling into the same hash
bucket.

In addition to supporting strings with null
characters, this version of the symbol table
handles conflicts more efficiently by skipping
over symbols with a different length.